gpt_telegram_bot/config/settings.py
import os
from dotenv import load_dotenv
from pydantic import BaseSettings

load_dotenv()

class Settings(BaseSettings):
    # Telegram
    telegram_token: str = os.getenv("TELEGRAM_TOKEN")
    
    # OpenAI
    openai_api_key: str = os.getenv("OPENAI_API_KEY")
    openai_base_url: str = os.getenv("OPENAI_BASE_URL", "https://api.openai.com/v1")
    
    # Database
    database_url: str = os.getenv("DATABASE_URL", "sqlite:///bot.db")
    
    # Redis
    redis_url: str = os.getenv("REDIS_URL", "redis://localhost:6379")
    
    class Config:
        env_file = ".env"

_settings = None

def get_settings() -> Settings:
    global _settings
    if _settings is None:
        _settings = Settings()
    return _settings 