requirements.txt
# GPT Telegram Bot Requirements

# Core packages
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL
# pymongo>=4.0.0  # Uncomment for MongoDB
# mysqlclient>=2.1.0  # Uncomment for MySQL

# Image processing
Pillow>=10.0.0
requests>=2.31.0

# Configuration and validation
pydantic>=2.0.0

# Logging
structlog>=23.0.0

# Caching and sessions
redis>=5.0.0

# HTTP client (optional, for advanced features)
aiohttp>=3.8.0

# Async support
asyncio-mqtt>=0.16.0  # Optional: for message queuing 